+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `excludedAttachmentElementIds` to `FormSubmissionEventConfiguration`
+- `ExecuteWorkflowEvent` to `FormsClient` class
 
 ## [6.0.0] - 2024-06-04